Make Hermes feed-event regression test compile and run; address review feedback
The FeedEventClassificationTests regression test never compiled on CI.
`CMUXCLI.classifyFeedEvent` lives in the `cmux_cli` executable module, but the
test (hosted by `cmux.app`) imported `cmux_DEV` and could neither see nor link
the CLI tool's symbols. The `tests` job was failing at the cmuxTests *build*
step (`cannot find 'CMUXCLI' in scope`) before any test ran — so the PR's
two-commit red/green never actually validated the fix.

Extract the pure classification logic into `CLI/FeedEventClassifier.swift`
(`struct FeedEventClassifier`, single source of truth), compiled into both the
`cmux-cli` and `cmuxTests` targets. The decision is now a genuine compiled unit
test: 10 tests run and pass. `@testable import cmux_cli` is not an option — the
app-hosted test bundle cannot link the CLI executable's symbols (the same
reason the existing CLI test uses a subprocess).

Behavior is unchanged for production payloads. Also addresses review feedback
folded into the extracted file:
- `FeedEventSemantic` is now `private` (Greptile P2).
- `ExitPlanMode`/`AskUserQuestion` routing deduplicated via a shared
  `dedicatedApprovalEvent(for:)` helper (CodeRabbit).
- Added coverage for Codex `beforeShellExecution` (telemetry) and generic-agent
  dedicated-approval-tool routing through the `.toolStartMaybeApproval` branch.
