+++ CLI/cmux.swift
@@ -30042,7 +30042,7 @@ export default function cmuxPiSessionExtension(pi: ExtensionAPI) {
         // Decide whether this event is Feed-actionable. Non-actionable
         // events are forwarded as telemetry (non-blocking) and exit `{}`
         // so the agent proceeds without a decision.
-        let (hookEventName, isActionable) = Self.classifyFeedEvent(
+        let (hookEventName, isActionable) = FeedEventClassifier.classify(
             source: source,
             event: rawEvent,
             toolName: toolName
@@ -30158,229 +30158,6 @@ export default function cmuxPiSessionExtension(pi: ExtensionAPI) {
         }
         print("{}")
     }
-
-    /// Classifies a raw agent hook event into our wire `hook_event_name`
-    /// plus an `isActionable` flag that drives whether the Feed bridge
-    /// blocks waiting for a user decision (and whether `FeedCoordinator`
-    /// posts a "needs approval" notification).
-    ///
-    /// The mapping is driven by an explicit, typed registry
-    /// (`feedEventSemantic`) keyed on `(source, event)` rather than by
-    /// pattern-matching raw event-name strings. Notification eligibility
-    /// is derived only from the resolved ``FeedEventSemantic``, so a
-    /// tool-*starting* lifecycle event can never be mistaken for an
-    /// approval request — and unknown / future event names default to
-    /// non-actionable telemetry that never notifies.
-    static func classifyFeedEvent(
-        source: String,
-        event: String,
-        toolName: String
-    ) -> (String, Bool) {
-        let semantic = feedEventSemantic(source: source, event: event)
-        return wireMapping(for: semantic, toolName: toolName)
-    }
-
-    /// User-attention semantic of a hook/feed event, independent of the
-    /// agent-specific raw event name. Notifications and blocking waits are
-    /// keyed off this — never off raw event-name string matching — so the
-    /// same misclassification cannot recur as new event names are added.
-    private enum FeedEventSemantic {
-        /// A real approval is pending; the user must approve/deny. Drives
-        /// the blocking Feed wait and the "needs approval" notification.
-        /// Resolved against the tool name so Claude's `ExitPlanMode` /
-        /// `AskUserQuestion` approvals route to their dedicated kinds.
-        case approvalRequest
-        /// A tool is about to run but no approval is pending. Telemetry
-        /// only. Used by agents that expose a *separate* approval event
-        /// (Claude, Codex, Hermes) so their pre-tool hook never escalates.
-        case toolStart
-        /// A tool is about to run and the agent has *no* dedicated approval
-        /// event, so a side-effecting tool is escalated to an approval and
-        /// read-only tools stay telemetry. Resolved against the tool name.
-        case toolStartMaybeApproval
-        /// A tool finished. Telemetry only.
-        case toolEnd
-        /// A new turn / prompt started. Telemetry only.
-        case promptSubmit
-        /// The agent finished responding. Telemetry only.
-        case response
-        /// A subagent finished responding. Telemetry only.
-        case subagentResponse
-        case sessionStart
-        case sessionEnd
-        /// A generic status/notification event. Telemetry only — real
-        /// approval banners for these agents fire through the dedicated
-        /// `notification` hook subcommand, not the feed path.
-        case statusNotification
-        /// Unknown / unregistered event. Safe default: telemetry only,
-        /// never actionable, never notifies.
-        case unknown
-    }
-
-    /// Resolves the semantic for a `(source, event)` pair. A registered
-    /// source uses its own table (unmatched events fall to ``unknown``);
-    /// unregistered sources use the generic table.
-    private static func feedEventSemantic(
-        source: String,
-        event: String
-    ) -> FeedEventSemantic {
-        let table = feedEventSemanticRegistry[source] ?? genericFeedEventSemantics
-        return table[event] ?? .unknown
-    }
-
-    /// Tool names that carry their own dedicated approval wire event rather
-    /// than the generic `PermissionRequest`. Returns the actionable wire
-    /// mapping for such a tool, or `nil` for ordinary tools.
-    private static func dedicatedApprovalEvent(for toolName: String) -> (String, Bool)? {
-        switch toolName {
-        case "ExitPlanMode": return ("ExitPlanMode", true)
-        case "AskUserQuestion": return ("AskUserQuestion", true)
-        default: return nil
-        }
-    }
-
-    /// Maps a resolved semantic to the wire `hook_event_name` plus the
-    /// `isActionable` flag, using `toolName` for the two tool-dependent
-    /// semantics.
-    private static func wireMapping(
-        for semantic: FeedEventSemantic,
-        toolName: String
-    ) -> (String, Bool) {
-        switch semantic {
-        case .approvalRequest:
-            return dedicatedApprovalEvent(for: toolName) ?? ("PermissionRequest", true)
-        case .toolStartMaybeApproval:
-            if let dedicated = dedicatedApprovalEvent(for: toolName) {
-                return dedicated
-            }
-            // Any tool that can mutate the environment surfaces as a
-            // permission request so the user can approve/deny from the
-            // Feed sidebar. Read-only tools stay non-actionable
-            // telemetry so we don't flood the Actionable view.
-            if Self.sideEffectingTools.contains(toolName) {
-                return ("PermissionRequest", true)
-            }
-            return ("PreToolUse", false)
-        case .toolStart:
-            return ("PreToolUse", false)
-        case .toolEnd:
-            return ("PostToolUse", false)
-        case .promptSubmit:
-            return ("UserPromptSubmit", false)
-        case .response:
-            return ("Stop", false)
-        case .subagentResponse:
-            return ("SubagentStop", false)
-        case .sessionStart:
-            return ("SessionStart", false)
-        case .sessionEnd:
-            return ("SessionEnd", false)
-        case .statusNotification:
-            return ("Notification", false)
-        case .unknown:
-            // Safe default: telemetry, no approval, no notification.
-            return ("PreToolUse", false)
-        }
-    }
-
-    /// Per-agent event-semantic tables. Each entry is the source of truth
-    /// for that agent's `(event) -> semantic` mapping; events absent here
-    /// resolve to ``FeedEventSemantic/unknown``.
-    ///
-    /// The key distinction the registry encodes: agents with a *dedicated*
-    /// approval event (Claude `PermissionRequest`, Codex `PermissionRequest`,
-    /// Hermes `pre_approval_request`) classify their pre-tool event as
-    /// ``FeedEventSemantic/toolStart`` (always telemetry). Agents whose only
-    /// signal is the pre-tool event (gemini, copilot, …, handled by
-    /// ``genericFeedEventSemantics``) use
-    /// ``FeedEventSemantic/toolStartMaybeApproval`` so side-effecting tools
-    /// still escalate. Conflating the two is the bug behind #4985.
-    private static let feedEventSemanticRegistry: [String: [String: FeedEventSemantic]] = [
-        "claude": [
-            "PermissionRequest": .approvalRequest,
-            "PreToolUse": .toolStart,
-            "PostToolUse": .toolEnd,
-            "UserPromptSubmit": .promptSubmit,
-            "SessionStart": .sessionStart,
-            "SessionEnd": .sessionEnd,
-            "Stop": .response,
-            "SubagentStop": .subagentResponse,
-            "Notification": .statusNotification,
-        ],
-        "codex": [
-            "PermissionRequest": .approvalRequest,
-            "PreToolUse": .toolStart,
-            "beforeShellExecution": .toolStart,
-            "PostToolUse": .toolEnd,
-            "UserPromptSubmit": .promptSubmit,
-            "SessionStart": .sessionStart,
-            "SessionEnd": .sessionEnd,
-            "Stop": .response,
-            "SubagentStop": .subagentResponse,
-            "Notification": .statusNotification,
-        ],
-        "hermes-agent": [
-            // `pre_tool_call` is a tool *starting* — Hermes raises a
-            // separate `pre_approval_request` for real approvals, so this
-            // must stay telemetry even for side-effecting tools (#4985).
-            "pre_tool_call": .toolStart,
-            "post_tool_call": .toolEnd,
-            // The approval banner for Hermes fires through the dedicated
-            // `notification` hook subcommand; on the feed path this stays a
-            // non-blocking notification to avoid a duplicate banner.
-            "pre_approval_request": .statusNotification,
-            "post_approval_response": .statusNotification,
-            "pre_llm_call": .promptSubmit,
-            "post_llm_call": .response,
-            "on_session_start": .sessionStart,
-            "on_session_reset": .sessionStart,
-            "on_session_end": .sessionEnd,
-            "on_session_finalize": .sessionEnd,
-        ],
-    ]
-
-    /// Fallback table for agents without a dedicated entry in
-    /// ``feedEventSemanticRegistry``. These agents expose only a pre-tool
-    /// event, so it carries ``FeedEventSemantic/toolStartMaybeApproval``.
-    private static let genericFeedEventSemantics: [String: FeedEventSemantic] = [
-        "PreToolUse": .toolStartMaybeApproval,
-        "beforeShellExecution": .toolStartMaybeApproval,
-        "PermissionRequest": .approvalRequest,
-        "PostToolUse": .toolEnd,
-        "UserPromptSubmit": .promptSubmit,
-        "SessionStart": .sessionStart,
-        "SessionEnd": .sessionEnd,
-        "Stop": .response,
-        "SubagentStop": .subagentResponse,
-        "Notification": .statusNotification,
-    ]
-
-    /// Tools that mutate state and deserve a user-visible approve/
-    /// deny prompt in Feed. Keyed on the canonical tool names Claude,
-    /// Codex, and similar agents emit. Read-only tools (Read, Grep,
-    /// Glob, Task, WebFetch, WebSearch, LS, TodoWrite, …) are
-    /// intentionally excluded.
-    private static let sideEffectingTools: Set<String> = [
-        "Bash",
-        "Write",
-        "Edit",
-        "MultiEdit",
-        "NotebookEdit",
-        "apply_patch",   // Codex
-        "shell",         // Codex / other agents
-        "terminal",      // Hermes Agent
-        "run_command",   // Antigravity
-        "write_to_file",
-        "replace_file_content",
-        "multi_replace_file_content",
-        "manage_task",
-        "schedule",
-        "ask_permission",
-        "invoke_subagent",
-        "define_subagent",
-        "manage_subagents",
-        "generate_image",
-    ]
 
     private static let skipInterviewAndPlanAnswer = "Skip interview and plan immediately"
